+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/ldap/TrustingSslSocketFactory.java
@@ -1,0 +1,67 @@
+package com.walmartlabs.concord.server.security.ldap;
+
+import com.google.common.base.Throwables;
+
+import javax.net.SocketFactory;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509TrustManager;
+import java.io.IOException;
+import java.net.InetAddress;
+import java.net.Socket;
+import java.net.UnknownHostException;
+import java.security.SecureRandom;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+
+public class TrustingSslSocketFactory extends SocketFactory {
+
+    private final SSLSocketFactory delegate;
+
+    public static synchronized SocketFactory getDefault() {
+        return new TrustingSslSocketFactory();
+    }
+
+    public TrustingSslSocketFactory() {
+        try {
+            SSLContext ctx = SSLContext.getInstance("TLS");
+            ctx.init(null, new TrustManager[]{new DummyTrustmanager()}, new SecureRandom());
+            delegate = ctx.getSocketFactory();
+        } catch (Exception e) {
+            throw Throwables.propagate(e);
+        }
+    }
+
+    @Override
+    public Socket createSocket(String s, int i) throws IOException, UnknownHostException {
+        return delegate.createSocket(s, i);
+    }
+
+    @Override
+    public Socket createSocket(String s, int i, InetAddress inetAddress, int i1) throws IOException, UnknownHostException {
+        return delegate.createSocket(s, i, inetAddress, i1);
+    }
+
+    @Override
+    public Socket createSocket(InetAddress inetAddress, int i) throws IOException {
+        return delegate.createSocket(inetAddress, i);
+    }
+
+    @Override
+    public Socket createSocket(InetAddress inetAddress, int i, InetAddress inetAddress1, int i1) throws IOException {
+        return delegate.createSocket(inetAddress, i, inetAddress1, i1);
+    }
+
+    private static class DummyTrustmanager implements X509TrustManager {
+        public void checkClientTrusted(X509Certificate[] xcs, String string) throws CertificateException {
+        }
+
+        public void checkServerTrusted(X509Certificate[] xcs, String string) throws CertificateException {
+        }
+
+        public X509Certificate[] getAcceptedIssuers() {
+            return new java.security.cert.X509Certificate[0];
+        }
+    }
+}